Decouple match admin handler from oRPC base to fix slow test cold-start
